Refactored

+++ src/main/java/pl/kosiorski/controller/CategoryController.java
@@ -5,8 +5,6 @@ import org.springframework.web.bind.annotation.*;
 import pl.kosiorski.dto.CategoryDto;
 import pl.kosiorski.exception.NoAuthorizationException;
 import pl.kosiorski.model.Category;
-import pl.kosiorski.model.User;
-import pl.kosiorski.repository.UserRepository;
 import pl.kosiorski.service.AuthService;
 import pl.kosiorski.service.CategoryService;
 import pl.kosiorski.service.UserService;
@@ -24,55 +22,84 @@ public class CategoryController {
 
   @Autowired
   public CategoryController(
-      CategoryService categoryService,
-      AuthService authService,
-      UserRepository userRepository,
-      UserService userService) {
+      CategoryService categoryService, AuthService authService, UserService userService) {
     this.categoryService = categoryService;
     this.authService = authService;
     this.userService = userService;
   }
 
   @GetMapping("")
-  public List<Category> getAll(@RequestHeader("Authorization") String token)
+  public List<CategoryDto> getAll(@RequestHeader("Authorization") String token)
       throws NoAuthorizationException {
-
-    List<Category> allByUserId;
-
-    try {
-      allByUserId = categoryService.findAllByUserId(userService.findByToken(token).getId());
-    } catch (Exception e) {
-      return null;
-    }
 
     if (authService.valid(token)) {
 
+      List<CategoryDto> allByUserId = categoryService.findAllByUserId(userService.findByToken(token).getId());
+      userService.generateNewToken(token);
       return allByUserId;
 
-    } else return null;
-
-
+    } else {
+      throw new NoAuthorizationException("You dont have authorization, try to login");
+    }
   }
 
   @GetMapping("/{id}")
-  public CategoryDto getOne(@PathVariable Long id) {
-    return categoryService.findById(id);
+  public CategoryDto getOne(@PathVariable Long id, @RequestHeader("Authorization") String token)
+      throws NoAuthorizationException {
+
+    if (authService.valid(token)) {
+
+      userService.generateNewToken(token);
+      return categoryService.findById(id);
+
+    } else {
+      throw new NoAuthorizationException("You dont have authorization, try to login");
+    }
   }
 
   @PostMapping("")
-  public CategoryDto save(@RequestBody @Valid CategoryDto CategoryDto) {
+  public CategoryDto save(
+      @RequestBody @Valid Category category, @RequestHeader("Authorization") String token)
+      throws NoAuthorizationException {
 
-    categoryService.save(CategoryDto);
-    return CategoryDto;
+    if (authService.valid(token)) {
+
+      categoryService.save(category, token);
+      userService.generateNewToken(token);
+      return category.toCategoryDto();
+
+    } else {
+      throw new NoAuthorizationException("You dont have authorization, try to login");
+    }
   }
 
   @DeleteMapping("/{id}")
-  public String delete(@PathVariable Long id) {
-    return categoryService.removeById(id);
+  public String delete(@PathVariable Long id, @RequestHeader("Authorization") String token)
+      throws NoAuthorizationException {
+    if (authService.valid(token)) {
+
+      userService.generateNewToken(token);
+      return categoryService.removeById(id);
+
+    } else {
+      throw new NoAuthorizationException("You dont have authorization, try to login");
+    }
   }
 
   @PutMapping("/{id}")
-  public CategoryDto updade(@PathVariable Long id, @Valid @RequestBody CategoryDto CategoryDto) {
-    return categoryService.updateById(id, CategoryDto);
+  public CategoryDto updade(
+      @PathVariable Long id,
+      @Valid @RequestBody Category category,
+      @RequestHeader("Authorization") String token)
+      throws NoAuthorizationException {
+
+    if (authService.valid(token)) {
+
+      userService.generateNewToken(token);
+      return categoryService.updateById(id, category);
+
+    } else {
+      throw new NoAuthorizationException("You dont have authorization, try to login");
+    }
   }
 }
